\name{da_report_conversion}
\alias{da_report_conversion}
\title{da_report_conversion}
\usage{
  trends_traffic(name=myfile)
  trends_traffic(myfile)
}
\description{
  DA Weekly Report - RAW-conversion 형식에 맞게 데이터 전처리를 도와주는 함수입니다.
}

\arguments{
  \item{name}{전처리된 파일을 로컬 컴퓨터에 저장할 때 파일명으로 사용되며, 반드시 입력해야 하는 필수 파라미터입니다.}
}
